Spostati alcuni metodi dei voti, Manuel Baldoni ha terminato la gestione delle classi (Pochi metodi).
